% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productsv3.R
\name{hs_create_product_association}
\alias{hs_create_product_association}
\title{Product create association (single)}
\usage{
hs_create_product_association(
  identifier,
  association_objecttype,
  association_identifier,
  association_type,
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)
}
\arguments{
\item{identifier}{The object identifier (either \code{vid} or \code{id}) of the related object.}

\item{association_objecttype}{The object identifier (\code{objectTypeId}) or unique name of the object
to associate with.}

\item{association_identifier}{The object identifier (either \code{vid} or \code{id}) to associate the main object to.}

\item{association_type}{String name of the association type.
Follows the \code{object_to_object} string format. A full list is
not available, but the \href{https://legacydocs.hubspot.com/docs/methods/crm-associations/crm-associations-overview}{CRM associations API}
table provides most association types. Custom objects create
new association type names as can be found in the \href{https://developers.hubspot.com/docs/api/crm/crm-custom-objects}{CRM Custom Objects API}.}

\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}
}
\value{
List with association data or \code{NULL} in case
no object was found (error 404), or \code{NA} if there was an
internal server error (error 500). All server error codes will display a warning.
}
\description{
Associate a product with another object.
}
\examples{
\dontrun{
associations <- hs_create_product_association(
  identifier = 1,
  association_objecttype = "contact",
  association_identifier = 25,
  association_type = "product_to_contact"
)
}
}
