% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects_mapper.R, R/objectsv3.R
\name{hs_create_company}
\alias{hs_create_company}
\alias{hs_create_contact}
\alias{hs_create_deal}
\alias{hs_create_line_item}
\alias{hs_create_product}
\alias{hs_create_ticket}
\alias{hs_create_object}
\title{Object create}
\usage{
hs_create_company(
  properties,
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)

hs_create_contact(
  properties,
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)

hs_create_deal(
  properties,
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)

hs_create_line_item(
  properties,
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)

hs_create_product(
  properties,
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)

hs_create_ticket(
  properties,
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)

hs_create_object(
  objecttype,
  properties,
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)
}
\arguments{
\item{properties}{Named list containing the properties to create or update.
To create or update multiple objects, each named list of properties needs
to be within a list (meaning, a list of named list with properties).}

\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}

\item{objecttype}{String name of the main object type.
Can be any of the default HubSpot objects (\code{contacts}, \code{companies}, \code{deals},
\code{tickets}, \code{line_items}, \code{products}, \code{quotes}) or a custom created object.}
}
\value{
List with created or updated object data (similar as \code{hs_get_\emph{xxx}()}).
The order of this list will NOT be the same as how you provided it into the function!
}
\description{
Create one or more objects with the given properties and return a copy of the object(s), including identifier(s).

The functions take into account batch limits for the API. This limit
is 100 objects per batch API call, except for the 'contacts' object,
where the limit is 10 objects per batch API call.
}
\examples{
\dontrun{
oneproduct <- hs_create_object(
  objecttype = "products",
  properties = list(name="New product",
                    description="Awesome product for everyone")
)
twoproducts <- hs_create_object(
  objecttype = "products",
  properties = list(list(name="First product",
                         description="Awesome product for everyone"),
                    list(name="Second product",
                         description="Even better product for everyone")
                   )
)
}
}
