% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{hs_create_schema_association}
\alias{hs_create_schema_association}
\title{Schema create association}
\usage{
hs_create_schema_association(
  object,
  association_objecttype,
  association_type = NULL,
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)
}
\arguments{
\item{object}{The object identifier (\code{objectTypeId}) or unique name of the object.}

\item{association_objecttype}{The object identifier (\code{objectTypeId}) or unique name of the object
to associate with. In case the name is provided, it is always the singular version of the
HubSpot object (contrary to the \code{objecttype} property).}

\item{association_type}{Unique string name for the new association type.
Common association types follow the \code{"object_to_object"} string format.
This parameter allows to create a custom association type name, but it is
not recommended to do this. By not providing this parameter, it will
automatically create the \code{"object_to_object"} name.

For example, a new association between a Car object and a Contact object
will be named \code{"car_to_contact"}.}

\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}
}
\value{
List with association data or \code{NULL} in case
no object was found (error 404), or \code{NA} if there was an
internal server error (error 500). All server error codes will display a warning.

The returned list of association data won't always be the same length
as the provided identifiers. Identifiers that do not have associations
are excluded from the returned list.
}
\description{
Associate a schema with another object.
}
\examples{
\dontrun{
association <- hs_create_schema_association(
  object = "car",
  association_objecttype = "contact"
)
}
}
