% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{hs_delete_schema_association}
\alias{hs_delete_schema_association}
\title{schema delete association}
\usage{
hs_delete_schema_association(
  object,
  association_objecttype,
  apikey = hubspot_key_get()
)
}
\arguments{
\item{object}{The object identifier (\code{objectTypeId}) or unique name of the object.}

\item{association_objecttype}{The object identifier (\code{objectTypeId}) or unique name of the object
to associate with.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}

\item{association_identifier}{The object identifier (either \code{vid} or \code{id}) to associate the main object to.}
}
\value{
This function always returns the same, an empty \code{raw} object.
If an incorrect input is provided, no warning or error will be given.
Obtain the association using the get association function (function \code{hs_get_\emph{xxx}_association})
with the same input to this function to see if it was properly deleted
(the returning association list would not have the \code{association_identifier} listed anymore).
}
\description{
Remove an association between the schema and another object.
}
\examples{
\dontrun{
result <- hs_delete_schema_association(
  identifier = 1,
  association_objecttype = "contact",
  association_identifier = 25,
  association_type = "schema_to_contact"
)
}
}
