% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productsv3.R
\name{hs_delete_product}
\alias{hs_delete_product}
\title{Product delete}
\usage{
hs_delete_product(
  identifiers,
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)
}
\arguments{
\item{identifiers}{One or more object identifiers (either \code{vid} or \code{id}) of the related object(s).}

\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}
}
\value{
This function always returns the same, an empty \code{raw} object.
If an incorrect identifier is provided, no warning or error will be given.

To see if an object is properly deleted, use \code{hs_get_\emph{xxx}} to obtain
the object with the same identifier (it should return \code{NULL}),
or use \code{hs_get_\emph{xxx}_list} to obtain multiple objects
with the same identifiers (it should return an empty list).
}
\description{
Delete one or multiple products.
}
\examples{
\dontrun{
oneproduct <- hs_delete_product(
  identifiers = 1
)
twoproducts <- hs_delete_product(
  identifiers = c(1,2)
)
}
}
