% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{hs_get_schema_list}
\alias{hs_get_schema_list}
\alias{hs_get_schema}
\title{Schema get one or a list}
\usage{
hs_get_schema_list(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)

hs_get_schema(
  object,
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}

\item{object}{The object identifier (\code{objectTypeId}) or unique name of the object.}
}
\value{
List with schema data, \code{NULL} in case
no object was found (error 404), or \code{NA} if there was an
internal server error (error 500). All server error codes will display a warning.
}
\description{
Get a list of custom schemas, or get a single one by name or identifier.
}
\examples{
schemas <- hs_get_schema_list()
oneschema <- hs_get_schema(
  object = "contact"
)
}
