% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productsv3.R
\name{hs_get_product}
\alias{hs_get_product}
\title{Product get by id}
\usage{
hs_get_product(
  identifiers,
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get(),
  properties = hs_object_properties_tidy("products",
    hs_object_properties_raw("products", token_path = token_path, apikey = apikey)),
  max_properties = 100
)
}
\arguments{
\item{identifiers}{One or more object identifiers (either \code{vid} or \code{id}) of the related object(s).}

\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}

\item{properties}{Character vector of properties to request}

\item{max_properties}{Avoid URLs that are too long, limit the number of
properties returned, if required.}
}
\value{
List with object data, \code{NULL} in case
no object was found (error 404), or \code{NA} if there was an
internal server error (error 500). All server error codes will display a warning.
}
\description{
Get one or more products.
}
\examples{
\dontrun{
oneproduct <- hs_get_product(
  identifiers = 1
)
twoproducts <- hs_get_product(
  identifiers = c(1,2)
)
}
}
