% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/account_details.R
\name{hs_account_details}
\alias{hs_account_details}
\title{Get account details}
\usage{
hs_account_details(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}
}
\description{
Get the account details for the API key or OAuth token.
}
\examples{
account <- hs_account_details()
}
