% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\name{hs_build_property}
\alias{hs_build_property}
\title{Property create}
\usage{
hs_build_property(
  name,
  label,
  type,
  field_type = NULL,
  group_name = NULL,
  description = NULL,
  option_labels = NULL,
  option_values = NULL,
  display_order = -1,
  has_unique_value = FALSE,
  hidden = FALSE,
  mainpropertylist = NULL
)
}
\arguments{
\item{name}{The internal property name, which must be used when referencing the property from the API.
Cannot contain spaces and is commonly lowercase letters and underscores only.}

\item{label}{A human-readable property label that will be shown in HubSpot. Start with capital letter.}

\item{type}{The type of property to create. Must be one of \code{enumeration}, \code{date}, \code{datetime}, \code{string}, \code{number} or \code{bool}.}

\item{field_type}{The method how the property appears in HubSpot or on a form.
If provided, it must match the provided \code{type}:
\item{\code{type}=\code{enumeration}}{This field can only be either of \code{select}, \code{checkbox}, or \code{radio}.}
\item{\code{type}=\code{date}}{This field can only be \code{date}.}
\item{\code{type}=\code{datetime}}{This field can only be \code{date}.}
\item{\code{type}=\code{string}}{This field can only be either of \code{text}, \code{textarea}, or \code{file}.}
\item{\code{type}=\code{number}}{This field can only be \code{number}.}
\item{\code{type}=\code{bool}}{This field can only be \code{booleancheckbox}.}
If this parameter is not provided (\code{NULL}) or the \code{type} only supports one \code{field_type},
it will take the first value mentioned in the list above.}

\item{group_name}{The name of the group this property belongs to.
The HubSpot API has no functionality to manage groups.
Default=\code{NULL}, which does not assign this property to a group.}

\item{description}{A description of the property that will be shown as help text in HubSpot.
Default=\code{NULL}, which leaves the description empty.}

\item{option_labels}{Vector of string labels for the available options
of an enumeration property (\code{type=enumeration}).
This field is only required for enumerated properties and is ignored for any other types.}

\item{option_values}{Vector of string values for the available options
of an enumeration property (\code{type=enumeration}). Must be of same length as \code{optionlabels}.
This field is only required for enumerated properties and is ignored for any other types.}

\item{display_order}{The order that this property should be displayed in the HubSpot user interface
relative to other properties for this object type. Properties are displayed in order starting
with the lowest positive integer value. A value of -1 (default) will cause the property to be displayed
after any positive values.}

\item{has_unique_value}{Logical. If \code{TRUE}, newly created objects must have this property
be different from existing objects with this property. If \code{FALSE}, duplicate values
of this property can occur (default).}

\item{hidden}{Logical. If \code{TRUE}, this property is invisible in the HubSpot user interface.
If \code{FALSE}, this property can be seen and used in the interface (default).}

\item{mainpropertylist}{Provide a list of already created properties to append
the newly created property to. Default=\code{NULL}.}
}
\value{
A list with the property settings, or a list of list with properties settings
when \code{mainpropertylist} is provided. Parameters that are left to its default
are not returned in the list.
}
\description{
Helper function to create a new object property with the right settings.
Used within the creation or editing of schemas of objects.
}
\examples{
newprop=hs_build_property(
  name = "make",
  label = "Make",
  type = "string",
  field_type = "text",
  description = "The manufacturer of a car"
)
}
