% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubspot-package.R
\docType{package}
\name{hubspot-package}
\alias{hubspot}
\alias{hubspot-package}
\title{hubspot}
\description{
Utilizing the \href{https://developers.hubspot.com/docs}{'Hubspot CRM' API} to
retrieve customer relationship data for reporting and analysis.
}
\author{
\strong{Maintainer}: Joost Kuckartz \email{jkuckartz1984@hotmail.com} [copyright holder]

Authors:
\itemize{
  \item Steph Locke \email{steph@itsalocke.com} [copyright holder]
  \item Maëlle Salmon \email{maelle.salmon@yahoo.se} (\href{https://orcid.org/0000-0002-2815-0399}{ORCID})
}

Other contributors:
\itemize{
  \item Jessie Geahlen \email{jgeahlen@gmail.com} [contributor]
  \item Locke Data (https://itsalocke.com) [funder]
  \item Michael W. Kearney \email{kearneymw@missouri.edu} (\href{https://orcid.org/0000-0002-0730-4694}{ORCID}) [contributor]
}

}
