% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_properties.R,
%   R/object_properties_mapper.R
\name{hs_object_properties_raw}
\alias{hs_object_properties_raw}
\alias{hs_object_properties_tidy}
\alias{hs_company_properties_raw}
\alias{hs_company_properties_tidy}
\alias{hs_contact_properties_raw}
\alias{hs_contact_properties_tidy}
\alias{hs_deal_properties_raw}
\alias{hs_deal_properties_tidy}
\alias{hs_product_properties_raw}
\alias{hs_product_properties_tidy}
\alias{hs_ticket_properties_raw}
\alias{hs_ticket_properties_tidy}
\title{Object properties endpoint (raw and tidy)}
\usage{
hs_object_properties_raw(
  objecttype,
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)

hs_object_properties_tidy(objecttype, object_properties = NULL, view = "names")

hs_company_properties_raw(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)

hs_company_properties_tidy(
  company_properties = hs_company_properties_raw(),
  view = "names"
)

hs_contact_properties_raw(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)

hs_contact_properties_tidy(
  contact_properties = hs_contact_properties_raw(),
  view = "names"
)

hs_deal_properties_raw(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)

hs_deal_properties_tidy(
  deal_properties = hs_deal_properties_raw(),
  view = "names"
)

hs_product_properties_raw(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)

hs_product_properties_tidy(
  product_properties = hs_product_properties_raw(),
  view = "names"
)

hs_ticket_properties_raw(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)

hs_ticket_properties_tidy(
  ticket_properties = hs_ticket_properties_raw(),
  view = "names"
)
}
\arguments{
\item{objecttype}{String name of the main object type.
Can be any of the default HubSpot objects (\code{contacts}, \code{companies}, \code{deals},
\code{tickets}, \code{line_items}, \code{products}, \code{quotes}) or a custom created object.}

\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}

\item{object_properties}{Object properties as returned by \code{hs_object_properties_raw()}}

\item{view}{A view (see Details)}
}
\value{
List of object properties (\code{hs_object_properties_raw()})

Something tidy (\code{hs_object_properties_tidy()})
}
\description{
Retrieve raw and tidy results from \href{https://legacydocs.hubspot.com/docs/methods/crm-properties/get-properties}{the object properties endpoint}.
}
\details{
Required scope(s) of the OAuth token: contacts.

Different \code{view} values and associated output.
\itemize{
\item "names": A character vector of the names of available object properties.
}
}
\examples{
properties <- hs_object_properties_raw("tickets")
property_names <- hs_object_properties_tidy("tickets", properties, view = "names")
}
