% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/products.R
\name{hs_products_raw}
\alias{hs_products_raw}
\alias{hs_products_tidy}
\title{Products endpoint (raw and tidy)}
\usage{
hs_products_raw(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get(),
  properties = hs_object_properties_tidy(hs_object_properties_raw("products",
    token_path = token_path, apikey = apikey)),
  property_history = "true",
  max_iter = 10,
  max_properties = 100,
  offsetvalue = 0
)

hs_products_tidy(products = hs_products_raw(), view = c("properties"))
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}

\item{properties}{Character vector of properties to request}

\item{property_history}{Whether version history for properties should be
returned}

\item{max_iter}{The API is limited to a certain number of responses by page, use \code{max_iter} to
limit how many pages of data will get returned}

\item{max_properties}{Avoid URLs that are too long, limit the number of
properties returned, if required.}

\item{offsetvalue}{The offset value to use if continuing a previous call. Defaults to 0.}

\item{products}{A products return from \code{hs_products_raw()}}

\item{view}{A view (see Details)}
}
\value{
List with products data (\code{hs_products_raw()})

A tibble with associated entities (\code{hs_products_tidy()})
}
\description{
Get raw and tidy results from the \href{https://legacydocs.hubspot.com/docs/methods/products/get-all-products}{products endpoint}.
}
\details{
Required scope(s) of the OAuth token: e-commerce

Different \code{view} values and associated output.
\itemize{
\item "properties": A tibble containing all properties of products
}
}
\examples{
products <- hs_products_raw(
  property_history = "false", max_iter = 1,
  max_properties = 10
)
products_properties <- hs_products_tidy(
  products,
  view = "properties"
)
}
