% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{hs_update_schema}
\alias{hs_update_schema}
\title{Schema update}
\usage{
hs_update_schema(
  object,
  display_properties = NULL,
  required_properties = NULL,
  searchable_properties = NULL,
  apikey = hubspot_key_get()
)
}
\arguments{
\item{object}{The object identifier (\code{objectTypeId}) or unique name of the object.}

\item{display_properties}{Vector of property names (must exist within the \code{properties} list)
that are used as display property (like \code{"firstname"}, \code{"lastname"}, \code{"email"}
and \code{"phone"} for a Contact object). The first property in this list will be
the primary display property, all remaining properties will be secondary display properties.
Display properties are commonly those properties that are mostly used or define the gross
of the object.}

\item{required_properties}{Vector of property names (must exist within the \code{properties} list)
that are required to be filled when creating new custom objects. Leave \code{NULL} (default) if no required
fields are needed (the Contact or Company object doesn't have any required properties for example).}

\item{searchable_properties}{Vector of property names (must exist within the \code{properties} list)
that can be searched within. Common to be text based fields, like \code{"firstname"} or \code{"lastname"}
for the Contact object. Leave \code{NULL} (default) if no searchable fields are needed.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}
}
\value{
List with created or updated schema data (similar as \code{hs_get_schema()}).
}
\description{
Update the details for an existing object schema.
Can only update the display, required, and searchable properties.
To add new properties to an object schema, use \code{\link{hs_create_property()}}.
To add new associations to an object schema, use \code{\link{hs_create_schema_association()}}.
}
\examples{
\dontrun{
schema <- hs_update_schema(
  object = "car",
  display_properties = c("make","model","year")
)
}
}
