% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productsv3.R
\name{hs_create_product}
\alias{hs_create_product}
\title{Product create}
\usage{
hs_create_product(
  properties,
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)
}
\arguments{
\item{properties}{Named list containing the properties to create or update.
To create or update multiple objects, each named list of properties needs
to be within a list (meaning, a list of named list with properties).}

\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}
}
\value{
List with created or updated object data (similar as \code{hs_get_\emph{xxx}()}).
}
\description{
Create one or more products with the given properties and return a copy of the object(s), including identifier(s).
}
\examples{
\dontrun{
oneproduct <- hs_create_product(
  properties = list(name="New product",
                    description="Awesome product for everyone")
)
twoproducts <- hs_create_product(
  properties = list(list(name="First product",
                         description="Awesome product for everyone"),
                    list(name="Second product",
                         description="Even better product for everyone")
                   )
)
}
}
