% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productsv3.R
\name{hs_delete_product_association}
\alias{hs_delete_product_association}
\title{Product delete association (single)}
\usage{
hs_delete_product_association(
  identifier,
  association_objecttype,
  association_identifier,
  association_type,
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)
}
\arguments{
\item{identifier}{The object identifier (either \code{vid} or \code{id}) of the related object.}

\item{association_objecttype}{The object identifier (\code{objectTypeId}) or unique name of the object
to associate with.}

\item{association_identifier}{The object identifier (either \code{vid} or \code{id}) to associate the main object to.}

\item{association_type}{String name of the association type.
Follows the \code{object_to_object} string format. A full list is
not available, but the \href{https://legacydocs.hubspot.com/docs/methods/crm-associations/crm-associations-overview}{CRM associations API}
table provides most association types. Custom objects create
new association type names as can be found in the \href{https://developers.hubspot.com/docs/api/crm/crm-custom-objects}{CRM Custom Objects API}.}

\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}
}
\value{
This function always returns the same, an empty \code{raw} object.
If an incorrect input is provided, no warning or error will be given.
Obtain the association using the get association function (function \code{hs_get_\emph{xxx}_association})
with the same input to this function to see if it was properly deleted
(the returning association list would not have the \code{association_identifier} listed anymore).
}
\description{
Remove an association between the product and another object.
}
\examples{
\dontrun{
result <- hs_delete_product_association(
  identifier = 1,
  association_objecttype = "contact",
  association_identifier = 25,
  association_type = "product_to_contact"
)
}
}
