% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{hs_delete_schema_association}
\alias{hs_delete_schema_association}
\title{schema delete association}
\usage{
hs_delete_schema_association(
  object,
  association_objecttype,
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)
}
\arguments{
\item{object}{The object identifier (\code{objectTypeId}) or unique name of the object.}

\item{association_objecttype}{The object identifier (\code{objectTypeId}) or unique name of the object
to associate with. In case the name is provided, it is always the singular version of the
HubSpot object (contrary to the \code{objecttype} property).}

\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}
}
\value{
This function always returns the same, an empty \code{raw} object.
If an incorrect input is provided, no warning or error will be given.
Obtain the association using the get association function (function \code{hs_get_xxx_association})
with the same input to this function to see if it was properly deleted
(the returning association list would not have the \code{identifiers_to} listed anymore).
}
\description{
Remove an association between the schema and another object.
}
\examples{
\dontrun{
result <- hs_delete_schema_association(
  object = "car",
  association_objecttype = "contact"
)
}
}
