% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productsv3.R
\name{hs_get_product_association}
\alias{hs_get_product_association}
\title{Product filter, sort, search
Product get associations (single)}
\usage{
hs_get_product_association(
  identifier,
  association_objecttype,
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get(),
  max_iter = 10,
  max_properties = 100,
  offsetvalue = 0
)
}
\arguments{
\item{identifier}{The object identifier (either \code{vid} or \code{id}) of the related object.}

\item{association_objecttype}{String name of the object type of the association.
Can be any of the default HubSpot objects (\code{contact}, \code{company}, \code{deal}, \code{ticket}, \code{line_item},
\code{product}, \code{quote}, or a custom created object).}

\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}

\item{max_iter}{The API is limited to a certain number of responses by page, use \code{max_iter} to
limit how many pages of data will get returned}

\item{max_properties}{Avoid URLs that are too long, limit the number of
properties returned, if required.}

\item{offsetvalue}{The offset value to use if continuing a previous call. Defaults to 0.}
}
\value{
List with association data or \code{NULL} in case
no object was found (error 404). A warning for error 404
will also be displayed.
}
\description{
Get a list of associations a single product has.
According to HubSpot, products don't have any (default) associations.
}
\examples{
\dontrun{
associations <- hs_get_product_association(
  max_iter = 1,
  association_objecttype = "contact"
)
}
}
