% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productsv3.R
\name{hs_get_product_list}
\alias{hs_get_product_list}
\title{Product get list}
\usage{
hs_get_product_list(
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get(),
  properties = hs_object_properties_tidy("products",
    hs_object_properties_raw("products", token_path = token_path, apikey = apikey)),
  max_iter = 10,
  max_properties = 100,
  offsetvalue = 0
)
}
\arguments{
\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}

\item{properties}{Character vector of properties to request}

\item{max_iter}{The API is limited to a certain number of responses by page, use \code{max_iter} to
limit how many pages of data will get returned}

\item{max_properties}{Avoid URLs that are too long, limit the number of
properties returned, if required.}

\item{offsetvalue}{The offset value to use if continuing a previous call. Defaults to 0.}
}
\value{
List with object data, \code{NULL} in case
no object was found (error 404), or \code{NA} if there was an
internal server error (error 500). All error codes will display a warning.
}
\description{
Get a list of products.
}
\examples{
products <- hs_get_product_list(
  max_iter = 1,
  max_properties = 10
)
}
