% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_properties.R
\name{hs_object_properties_raw}
\alias{hs_object_properties_raw}
\alias{hs_object_properties_tidy}
\title{Object properties endpoint (raw and tidy)}
\usage{
hs_object_properties_raw(
  objectname,
  token_path = hubspot_token_get(),
  apikey = hubspot_key_get()
)

hs_object_properties_tidy(objectname, object_properties = NULL, view = "names")
}
\arguments{
\item{objectname}{String name of the object. Can be one of \code{"tickets"}, \code{"products"} or \code{"line-items"}.}

\item{token_path}{Path to OAuth 2.0 token.
See \code{\link[=hubspot_token_create]{hubspot_token_create()}}.}

\item{apikey}{API key to work with Hubspot (only used if \code{token_path}
is \code{NULL}). See \code{\link[=hubspot_key_set]{hubspot_key_set()}}.}

\item{object_properties}{Object properties as returned by \code{hs_object_properties_raw()}}

\item{view}{A view (see Details)}
}
\value{
List of object properties (\code{hs_object_properties_raw()})

Something tidy (\code{hs_object_properties_tidy()})
}
\description{
Retrieve raw and tidy results from \href{https://legacydocs.hubspot.com/docs/methods/crm-properties/get-properties}{the object properties endpoint}.
}
\details{
Required scope(s) of the OAuth token: contacts.

Different \code{view} values and associated output.
\itemize{
\item "names": A character vector of the names of available object properties.
}
}
\examples{
properties <- hs_object_properties_raw("tickets")
property_names <- hs_object_properties_tidy("tickets", properties, view = "names")
}
